requirements.txt
streamlit
openai-whisper
nltk
reportlab